Unify PID type

+++ pkg/appolly/discover/container_updater.go
@@ -46,7 +46,7 @@ func updateLoop(
 				switch ev.Type {
 				case EventCreated:
 					log.Debug("adding process", "pid", ev.Obj.FileInfo.Pid)
-					db.AddProcess(uint32(ev.Obj.FileInfo.Pid))
+					db.AddProcess(ev.Obj.FileInfo.Pid)
 				case EventDeleted:
 					// we don't need to handle process deletion from here, as the Kubernetes informer will
 					// remove the process from the database when the Pod that contains it is deleted.
